+++ app/Http/Controllers/Admin/CategoryController.php
@@ -153,13 +153,18 @@ class CategoryController extends Controller
      */
     public function method(Request $request)
     {
-        if($request->option == 'delete')
+        if($request->option == 'delete'&& $request->checkbox != null)
         {
             return $this->destroy($request);
         }
-        elseif($request->option == 'activate')
+        elseif($request->option == 'activate'&& $request->checkbox != null)
         {
             return $this->activate($request); 
+        }
+        else
+        {
+            $request->session()->flash('fail', 'Hãy chọn tác vụ hoặc chọn bất kì 1 ô nào đó !!! ');
+            return redirect()->route('indexCategory');
         }
     }
     public function destroy(Request $request)
